+++ NEWS.md
@@ -8,6 +8,11 @@ averaged over the rolling window.
 - Fixed a bug where `transmute_weights()` and `factor_weights()` could return
 a result with a different length than `w`.
 
+- Added a new function `splice_index()` for splicing indexes calculated over
+a rolling window (this was previously sketched in an example).
+
+- `transmute_weights()` is now faster.
+
 ## Version 0.6.0
 
 - Bumped minimum version of R to at least 4.0.